size of integers in f77
Former-commit-id: 7e2ca413a7591173bd609ee3201eed82d72b7666

+++ api/f77/Testfile.f
@@ -3,7 +3,7 @@
       integer sf_gettype
       integer type, n1, i
       logical sf_histint
-      integer in, out, sf_input, sf_output
+      integer*8 in, out, sf_input, sf_output
       character*100 sf_histstring, label1
       real trace(100)
